+++ Application/translations/en/translations.php
@@ -25,6 +25,7 @@ return [
     'D3_TOTP_CANCEL_LOGIN'                            => 'Cancel login',
     'D3_TOTP_BREADCRUMB'                              => 'one-time password login',
     'D3_TOTP_ERROR_UNVALID'                           => 'The one-time password is invalid.',
+    'D3_TOTP_ERROR_REPLAY'                            => 'The one-time password cannot be used again. Please wait a moment.',
     'D3_TOTP_ACCOUNT'                                 => '2-factor authentication',
     'D3_TOTP_ACCOUNT_DESC'                            => 'Secure your account login with a second factor.',