+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [3.0.0.0](https://git.d3data.de/D3Public/oxtotp/compare/2.1.1.0...3.0.0.0) - 2026-06-XX
 ### Added
 - installable in OXID eShop 7.0 to 7.4
+- replay protection
 ### Changed
 - use separated session variable names between front and backend
 - use generated master key instead of static one